Support returning pandas dataframe from sql exec
If the user does not specify a output array and the query produces
results we now will return the results in a json array to the user. By
default this json bytes will be converted to a pandas dataframe.

The goal is to allow returning relatively smaller result sets, such as
an aggregate without having to store in a output array.