Add support for rendering visualization graph
Two tree formats are supported, one using pydot if the user has dot/graphviz available.
The second using a python method to compute the layout.

Graphs are build using plotly and can use normal figures or widgets if inside a notebook.

Graph figures can be auto-updated in realtime using callback on the status