Fix incorrect datetime ranges in UDF apply
If the user supplies a np.datetime64 start/end range with a different time resolution than the
underlying dimension in the schema, then currently we cast directly to int64 which will
not be correct when applied server-side. In this patch we add passthrough for the schema,
and then apply the corresponding dimension dtype before getting the raw int64 value.
Untested, pending fixes for unrelated issue server-side.