Copy signature of async array and UDF functions onto sync versions.
This change introduces `signature_of`, which copies the function
signature from a source to the decorated function, and applies it to
the synchronous functions in array.py and udf.py.

This avoids potential issues with parameter order or defaults falling
out of sync between the synchronous and the async versions, while still
allowing interactive users to access tab-completion. The docstrings
of the synchronous versions link directly to the async versions.

I have manually tested this both in IPython and Jupyter and tab
completion and other contextual help continues to work there.