Allow array UDFs to pass either a callable or a registered name.
This change allows users to pass their UDFs to TileDB Cloud as either
a callable or the name of a registered UDF. It changes the `func` param
from callable-only to callable-or-name, and deprecates the `name` param.

It also includes a suite of cleanups that were made along the way
to reduce duplication.