Add `replace_stored_params`, to support loading stored params.
This adds the `replace_stored_params` function, which uses the
infrastructure established in the previous change to load in data that
will be passed into the UDF environment as a stored parameter.

This genericizes the `_ResultReplacer` class with a `visit` function
that tells the replacer what to replace a parameter with. The visit
function can raise a `_ReplaceWith` exception to indicate that the
instance currently being visited should be replaced with the value
included in the exception. Otherwise, the `_ResultReplacer` proceeds
as normal, descending into instances (or not doing so, as appropriate).