Move decoding of results into classes in `results.py`.
This moves most decoding logic into Decoder classes, which are designed
to be easily pickleable for future use in stored parameters. It also
introduces an early version of stored-parameter decoding in
`dag/stored_params.py`, which will eventually be used (along with a
reworking of async operations and DAG `Node`s) to avoid sending
parameters back to the server.

There are also some changes to make errors more explicit:

- Rather than silently ignoring unknown result types, they are raised
  as errors.
- Results returned as Pandas are now guaranteed to be returned as a
  DataFrame, rather than “a DataFrame, or potentially some other type”.