Remove "fresh" dependency strategy from the matrix. (#172)
This reverts to essentially the version of the pipeline YAML file from
before we added the "fresh checkout" version. It seems like running too
many tests concurrently is causing us trouble where the server gives us
502s, so this is an attempt to reduce it. Limiting the number of runners
means that we would have to wait for two (or more) tests to complete
one after the other. In the context of a test pipeline, pinned
dependencies are more useful than a fresh install.