Further reduce autotest concurrency and add a nightly build.
This stops running automated on-push tests on Mac OS (for now).
To make up for this loss, it adds a nightly test pipeline that runs
our full matrix of tests, the cross product of:

- Mac OS and Linux
- Python 3.6 and 3.7
- requirements.txt and a fresh `pip install`

These are limited to one run at a time, since time is no object when
running tests at night.