Avoid replacing nodes when it is not needed.
This adds a pre-probing step to the replacement logic, to descend
through a data structure to see *if* any values need to be replaced
before actually constructing a replacement.

If neither the current value nor any of its children have any values
that need replacement, then we return the node unmodified.