Make the primary implementations of UDFs synchronous.
Currently, the canonical implementation of UDFs is to use the generated
API’s asynchronous interface, and the synchronous version wraps that
and gets the result of the `multiprocessing.pool.TaskResult` returned by
the generated code. This will not work for our plans to avoid
round-tripping all task results when running DAGs.

This change instead makes a new `whatever_base` as the canonical base
implementation of a UDF caller, and bases the existing `whatever` and
`whatever_async` functions off of them, maintaining compatibility.
Additionally, we manage our own thread pool for async requests, which
avoids us having to use multiprocessing’s poorly-supported `TaskResult`
interface and gives us better control over how many threads we run.
UDF request management is fully moved into the `Client` object.

Results are returned by a new `results.Response` class, which includes
the important `task_id` information that will be needed to avoid
round-tripping requests in the future, and a `results.AsyncResponse`
class also wraps (a `futures.Future` of) that `Response` to provide
an API that is 100% compatible for callers currently expecting a
`UDFResult` or `SQLResult`.

Additional changes:

- Creates exactly 1 `ApiClient` to better reuse network connections.
- Cleans up API instance creation.
- Adds types in some places.
- Cleans up imports in tests.