Add client support for stored parameters.
This finally completes all the wiring needed to get the server to load
stored parameters rather than having to serialize giant blobs both ways.

- Pulls decoders into `decoders.py`. This is needed to break a circular
  dependency and allow the abstract `Result` class to live in
  `results.py`.
- Adds an abstract `Result` class to `results.py` and makes
  `RemoteResult` and `LocalResult` both implementations of it.
  The `Result` is the thing that knows how to convert itself into
  a `StoredParam`.
- Adds `stored_params` plumbing into the various `whatever_base`
  functions so that we can pass stored parameters into them and have
  them be included in calls.

The major change is a substantial reworking of the Node class.
The functions that a Node actually executes return `Result` instances,
which are used to track the result and the necessary metadata to know
whether we are able to load them remotely.

A caller’s passed-in function is wrapped into a `LocalResult` instance.
`XXX_base` functions that make remote calls short-circuit this wrapping,
as they already return a `RemoteResult` (itself a `Result` subclass).
Instead of passing a stored-result function directly into the Executor,
we use a function that handles substitutions and retries, so that the
existing Future-base interface will continue to work.

Because SQL queries are not executed in the UDF environment,
all parameters need to be passed directly and we cannot use
stored parameters.