Update DAGs to not download intermediate results by default. (#200)
This enables callers to not actually download the results of a
server-side call with a new `_download_results` parameter to each of the
`whatever_base` functions. If set, it will tell the server not to send
the results of the function when it is complete; just to return headers.
(This is useful even *before* the server has support for not sending
results back—we will get the HTTP response and completely ignore
the contents, meaning that the dataset will not be parsed into memory.)

This requires restructuring of the Node class itself, but maintains
(much of) the same external interface.

The actual future used by the instance internally is privatized and now
wraps a `Result` rather than the object that the UDF call would have
produced. The (not-quite official, but still present) external-facing
`.future` is now a property that materializes a proxy `Future` that
receives the result of the actual Future.

Also updates dependencies and requirements txt files.