Enable keepalive pings on sockets to prevent connection drops.
For particularly long-running UDFs or SQL queries, the server may go
minutes or more without responding with any data back to the client.
This means that NATs or load balancers may drop the connection,
resulting in the loss of all the UDF's progress. This change enables
1-minutely TCP keepalive pings on sockets, so idle connections are
kept fresh and do not have an opportunity to go stale.

Short-lived connections and long-lived connections that do not go idle
due to ongoing transfers are unaffected.