Fix error with zero-length Arrow response bodies.
For situations where a UDF was successful but for whatever reason
the result returned was a zero-length body, we would mistakenly assume
that the body was not provided, rather than recognizing it as a valid
but zero-length body. We now distinguish between no body and a
zero-length body, and properly handle zero-length Arrow bodies.

We had previously tested against empty Arrow bodies with SQL queries.
SQL queries include a schema in their results, and thus have nonzero
response length. However, empty UDF responses which are written in Arrow
format would not have any write calls at all, meaning the response was
completely empty.

Also adds a test for this situation.