Allow `Delayed` task graphs to log themselves. (#238)
Previously, task graphs created through the `Delayed` API did not
properly log themselves with the task graph logging interface, because
they bypassed a bunch of the API machinery that set things up for that.
This change ensures that even `Delayed` task graphs can log themselves
by using information that is passed as part of the `Delayed` API.

Most of our tests (and possibly some users) use the `add_node` function
to add nodes to the graph. These are locally executed and we cannot add
our remote-calling parameters to the functions (as they are not our
functions which make requests to the server).