Add a `json_safe.Value` sentinel type for API serialization.
This avoids the API client from descending through values that we know
are already validly prepared for JSON serialization (i.e., only dicts,
lists, strings, and numbers) to look for its own `models.*` types.
For instance, currently if we have a call like

    call_some_api(ranges=[list, of, a, million, ints, ...])

the generated client will look inside each of `ranges` to see if
any of those million ints are in fact some object that it knows about.
But with this change, we can say:

    call_some_api(
        ranges=json_safe.Value([list, of, a, million, ints, ...]),
    )

and the generated client will just use that list and assume that it has
already been sanitized.