Allow nodes from old DAGs to be used in new DAGs.
Previously, if you ran a DAG either via DAG or Delayed and passed in a
Node from a previous graph, it would either not work at all or give you
a bizarre error message (like "graph does not have any root nodes").
With this change, you can pass Nodes from a completed DAG to a new DAG
and they will be used as a caller would expect as parameters.