Move non–result-specific code to a new `_common` module.
To reduce the pollution of the `_results` namespace, this moves common
stuff that is not result-specific into its own internal-only module.
Also adds and cleans up some generic types.

Most of what remains in `_results` is specific to resultset handling,
and much of it cannot be moved due to the need to maintain pickle
compatibility for v1 UDFs.