Report of cancelled DAG nodes and client-side failures.
This adds the ability for the client to report information about events
related to DAG nodes that only ever happened on the client side. For
graphs where we've already submitted logs to the server, we will now
report when a node completes on the client side, whether successfully
or through failure or cancellation.

This also has some minor implications for the way cancellation works:
Previously, if a task was already actively running, cancelling it would
mark it as cancelled immediately. However, as far as the actual code,
the task would still be running and would even be able to return its
result had it not been marked as cancelled. This change updates the way
cancellation works in the API to reflect the actual behavior of Python
futures: where once a task is running, it can't be cancelled and will
still return a value.