Clean up thread pool when a DAG is done.
This change cleans up the DAG thread pool when the DAG is complete,
so that it can release any thread resources that are still active.
It also cleans up the way we update status and check doneness:
rather than using the `done` method as a cudgel to cause the graph to
update its status, this change just has the graph update its own status
when it receives an update from a Node.

Also cleans up some synchronization in the tests to use Barriers rather
than an unintuitive use of Futures.