Only cancel descendants of failed DAG nodes (#274)
* Restructure Node to use a Future for all lifecycle management.

This eliminates a lot of sources of error by making the Future that
the Node uses be the single source of truth for what its status is.
All lifecycle commands and events affect it, and it in turn forwards
those events onto its result proxy and the downstream listeners.

* Only cancel dependencies when a DAG node fails.

This improves our handling of DAG failures to only cancel those nodes
that depend upon a failed (or cancelled) node when a node fails (or is
cancelled). Nodes independent of the failed (or cancelled) node will
continue to be scheduled and executed.

* Clean up thread pool when a DAG is done.

This change cleans up the DAG thread pool when the DAG is complete,
so that it can release any thread resources that are still active.
It also cleans up the way we update status and check doneness:
rather than using the `done` method as a cudgel to cause the graph to
update its status, this change just has the graph update its own status
when it receives an update from a Node.

Also cleans up some synchronization in the tests to use Barriers rather
than an unintuitive use of Futures.

* Deflake `test_simple_cloud_dag`.

Due to replication and synchronicity, `test_simple_cloud_dag`
occasionally sees incorrect results. Waiting for just a short time
before checking result status allows the update to be fully reflected.