Define the interfaces an Executor and a Node fulfill.
This creates the `executor` module. This specifies the interfaces that
any implementation of an Executor and its various Nodes will fulfill.
This base behavior will allow users to interact with multiple Executor
implementations in the same way. The initial implementation will be a
client-side executor, but in the future we may provide additional
implementations, like a server-side executor or other remote execution
model.