Create the initial framework for the client executor.
This is the very initial parts of the client-side executor, which
coordinates execution of task graphs on the client side. It consists of
the beginning of the Executor implementation and the abstract base class
for the client-side executor's Nodes.

Later commits will complete the implementation of the Executor as well
as implementing each type of Node itself.