Create the Array Node.
The array node is a virtual node which is used to set up the server-side
querying of a TileDB array to be passed as a parameter to a UDF.
It turns an ordinary UDF into an array UDF. The local executor needs to
read the inputs and create an array-read specification for the node's
output to be passed to the server. From there, the UDF runtime handles
the rest.