Add InputNode, the final node type.
The InputNode is the node that lets authors parameterize their graphs.
After putting it in, it works essentially like any other node, but it
is set by the user when they call `execute` on the graph. It can be used
as a parameter in any place where Nodes are accepted as params.