Keep interpreter alive while reporting graph status.
In situations where the user was executing a task graph and the program
finished immediately after that (e.g. `grf.wait()` is the last statement
in `main()`), the interpreter would often exit before the event loop
got the chance to report the graph's completion. This change ensures
that completion will be reported, but it still won't block the user's
other code in cases where there is more after `.wait()` (or equivalent).