Prepeare executor implementation for retry support.
This reworks the implementation of the primary executor to use a
condition variable, not an event, to signal the execution's completion.
This allows us to reset the state of the executor back to "in progress"
after it finishes for the first time.

Retries are not yet implemented, but this is part of the groundwork.