Allow callers to retry failed or cancelled task graph nodes.
This adds retry logic to the new task graph API. The `Executor` itself
is responsible for handling all retries, with a convenience method on
each `Node` to perform a retry if necessary.

Retries are handled on the same event loop as all other graph lifecycle
management operations, and are accomplished by resetting the state of
the to-be-retried node, re-walking the graph, and re-enqueuing all nodes
that can and should be re-run as needed.