Allow nodes of modern task graphs to be executed locally.
This is a first implementation of local execution for task graphs.
All local tasks are executed on a separate process with multiprocessing.

In the new executor, task graphs require their functions to be pickled.
For this reason it is impossible to directly share memory between the
process executing the task graph and the worker itself (e.g. you can’t
append to a list). Because there is no such shared memory, there is no
reason not to spawn separate processes for local work, particularly
when it might be complex and hog the GIL.

The main use case in our documentation for a local UDF is to interact
with local resources, like the file system. For this reason, there is
no functionality to automatically submit a UDF to the server if we are
unable to execute that UDF locally; instead we simply say that it cannot
be executed locally.

Improvements planned for follow-ons:

- Use a process pool, stored in the Executor, rather than creating
  a new process for each UDF call.
- Separate the count of UDFs that are allowed to be run in parallel
  on the server side from those on the client side (i.e. even if you're
  already running your maximum of 5 UDFs on the server, client-side
  execution has a separate quota pool).
- Add a way to (manually) allow local UDFs to be run remotely if they
  fail to run locally.