Wire DelayedGraph up to the executor.
This sets up the execution of task graphs built with Delayed.
When any node of a connected DelayedGraph is started, all the nodes
on that graph are built with a TaskGraphBuilder and executed with the
default executor. The entire execution workflow is handled by the
executor, and the DelayedGraph has no execution logic of its own.
The graph holds onto that executor and uses it to get references to the
execution nodes as they are needed, e.g. to get results, and manage
lifecycle events.

Nodes are extended to implement the Future-like interface with some
additional Delayed-specific lifecycle management methods. By itself,
this commit is a bit empty without the next one, which includes the
implementation of delayed UDFs...