Add the Array Node, which enables Array UDFs.
This adds and tests the Array node, so that users can actually start
doing interesting TileDB-specific things with their UDFs by passing in
their TileDB arrays as values.

This also streamlines the process of default-handling with `NOTHING`,
by making a central place to strip `NOTHING`s out of kwarg dictionaries.