Add encoding for slice data and numpy datetimes.
We will rely on the ability for non-Python code to decode Slices used
to specify query ranges in arrays.

For this reason, we need to encode both Slices and numpy datetimes
in a format that can be read universally. This adds the `__slice__`
and `__np_time__` formats into our JSON encoding, so that code in any
language can encode and decode data that translates into Python slices
and numpy datetimes.