Reimplement the Delayed API atop the new task graph API.
The existing Delayed API is an extremely thin layer on top of the old
DAG API, which is itself a relatively thin layer atop the various
`exec_something` functions. The new task graph API is neither of those
things. The implementation of the old one leaked a whole lot of
abstraction, and this new implementation represents a best-effort
attempt to match the behavior of the existing Delayed API atop the new
task graph API.

To the best of my knowledge and ability, existing code that uses
“standard” methods of the existing Delayed API should continue to work
unchanged. However, this breaks compatibility with essentially anything
that reaches into or otherwise depends upon the internals of the
existing Delayed objects, including any of the `Node` specifics or the
`delayed_obj.dag` DAG object. Common interactions where that takes place
have been replaced with methods on the common node type.

Unfortunately, this has to be one commit because, since the existing API
is one integrated package, it can only be replaced all at once.