Improve visualization graph layout by making nodes round.
The visualizer shells out to GraphViz to actually generate the layout
for the graph that appears in the Jupyter notebook. By default, it uses
the ID of the node (which is a UUID) as the label of the node, which
means that every node is a big wide oval. This makes for weirdness when
layout is done, because the big wide ovals do not look like the circles
that are actually used when rendering the graph.

This change explicitly clears the label and sets the height and width
of the graphviz nodes so that they are circles and should lay out
much more nicely.