Overwrite logic in notebook arrays (#339)
This PR will make easier the progress of [sc-20869], plus will enhance
the cloud codebase with some semantics considered in the past as well.
It introduces an Enum class OnExists for selecting upon uploading the
desirable logic. A prerequisite for overwriting logic is the existence
of the notebook at the first place.