Drop Python 3.6 from the build.
As of https://github.com/TileDB-Inc/TileDB-Py/pull/1300, Python 3.6
is no longer supported by the `tiledb` package (i.e., there is no
TileDB Python 3.6 wheel).  The wheel that we build for `tiledb-cloud`
is Python version–independent, so the package will remain installable
by Py3.6 users so long as the code uses no 3.7+ constructs (but we will
no longer make a special effort to retain this compatibility).