Add tool to rewrite API URLs and operation names.
Because our two APIs live at `/v1` and `/v2`, we ordinarily would not
be able to use the same API configuration instance for both of them.
`rewrite_urls.py` rewrites the URLs in the API document so that the
client generator can access them both from the same root, meaning that
we can use the same config implementations (as well as the rest of the
invariant API stuff) across both.