Regenerate v1 API with path prefixes.
This regenerates the TileDB Cloud v1 API, but adds the `/v1` prefixes
to each individual operation rather than having it in the config.
This also removes the logic which adds `/v1` to the path when it is
passed in. Everything continues to work exactly as it did before
on the client side.

Generated based on rev 6e1428ada5ab248135be3c066a7e60b27807874a.