Update the generator to generate v2 of the API.
This adds the v2 generation step, as well as adding override files.
Because the generated APIs use absolute references to certain types
(exceptions, etc.), we need to ensure that both copies of the API refer
to the exact same types and instances, so this ensures that the contents
of all common namespaces are duplicated across files.

The API is generated into `_common/v2_api` to make it explicit that
the raw API is really internal-only and users should just use the
helpers that we expose.