Don't use multiprocessing when running nodes locally. (#345)
Despite the potential advantages of running local nodes in multiple
threads from 008cf67248b28aa15a810827e6b610e4b2959ef2 when it was added,
there are difficult-to-diagnose issues with multiprocessing where it
seems like certain processes are never starting. For this reason, we're
dropping back to just executing everything on the same thread.