Always drain the HTTP connection when not preloading response. (#352)
When `preload_content` is set to False on a urllib3 HTTP request, the
connection is not automatically returned to the pool, and the contents
of the response body are not read. This means that, if the code does not
explicitly call `.read()`, control may be returned to the parent before
the server has "finalized" the HTTP response, which in our case means
the operation requested may not have fully completed.

For both of these reasons, we always need to manually return the
connection to its pool, and `utils.release_connection` does that for us.