Don't fail when we can't save the config file on login.
Users running in a read-only filesystem may not be able to write the
TileDB Cloud config file but this should not prevent login. Record
a warning so that the user can be alerted of the situation if it's not
intended, but allow execution to continue.

Fixes https://github.com/TileDB-Inc/TileDB-Cloud-Py/issues/354