Fix dependency recording for DelayedArrayUDFs. (#357)
Dependencies were resolved but not correctly recorded for
DelayedArrayUDF objects, which meant that attempting to use previous
nodes as parameters to array UDFs would fail. This fixes that.

Also includes:

Fix calling Delayed add_done_callback before a graph starts.

A missing `()` meant that if it were called before the graph was
started, the callback would not be added correctly.