Add tool to force function pickling by value.
When registering certain UDFs, like `trampoline.run_python_function`,
we want to ensure that they're registered by reference, not by value.
Removing the `__module__` from the function is the simplest way to
make that happen with the current version of cloudpickle.

This may be worth exporting in the future, or alternately even using it
in the default UDF execution/registration process, but until we have
a better idea of how to work with it, I'd like it to stay internal-only.