Run all i/o parts of the bioimaging ingestor in the graph.
Because we cannot (currently) provide access credentials to a UDF
executed directly from the client, we need to do all i/o related tasks
in the server-side task graph that we build. This pulls the initial
listing of input URIs into the graph itself and uses the parallelization
support ("output expansion") of server-side graphs to run the ingestion
in parallel.