Make 3 attempts to run Python tests before failing the build. (#430)
On the best of days, automated tests that access the network can be
flaky and prone to hiccups, and our tests here are no different.
For this reason, we want to retry them to avoid the hassle of repeatedly
manually retrying the tests waiting for them all to succeed. We use
pytest to rerun only failed tests.