Add and test codec for encoding Pandas via Arrow.
We currently serialize Pandas data using ordinary Python Pickles, but
doing so across Python versions and Pandas versions is extremely
unsupported.  We also would like to fully decouple the version of Pandas
between the client and the server, so that users who need Pandas 1.x can
use it, but users who need Pandas 2.0 can use it in UDFs, with no added
effort on either of their parts.

Fortunately, Apache Arrow already supports serializing Pandas data
in a cross-compatible format. Adding this new codec type will allow us
to store Pandas DataFrames in the "TileDB JSON" format, with encoding
and decoding being transparent to the user.

This change only adds the decoder; we need to wait for a release of this
to enable encoding it, so that we're sure that we can decode it when we
actually start *sending* Pandas DataFrames via TileDB JSON.