Use replacing visitor so we actually descend into batch args. (#499)
When server-side ("batch") DAG execution was added, the implementation
used custom code for finding parent nodes, rather than the existing
visitor system. This meant that if a Node was passed in not at the top
level, it would not work.

This change replaces it with a visitor-based system that allows users to
pass in nodes just as they would in client-side execution.