Update code generation to support v2 API returns.
When returning data from an API, the generated OpenAPI code looks it up
by name. This meant that, since we needed to use the same API client
types across both the v1 and v2 APIs, v2 API returns would attempt to
look up their named types in `tiledb.cloud.rest_api.models`, where they
would not exist.

We managed to get away with this until something in the v2 API changed
and was defined as returning an object instead of None.