Fix new test partitioning in GitHub Actions runs. (#580)
When adding a new test to the suite, it won't have timing information
on its initial run. However, once it runs once, it will have timings,
but adding its timings to the existing database may result in the
sharding getting changed. This means that if the test fails, the rerun
may not include the failing test, as seen in test logs:

https://github.com/TileDB-Inc/TileDB-Cloud-Py/actions/runs/9388190301/job/25852891056#step:11:160

    [pytest-split] Splitting tests with algorithm: least_duration
    [pytest-split] Running group 2/6 (estimated duration: 114.28s)

    collected 245 items / 184 deselected / 61 selected
    run-last-failure: 1 known failures not in selected tests

Instead, we ask pytest to run exclusively the tests that just failed
when doing a rerun, and don't re-shard at all.