Put urllib3.BaseHTTPResponse annotation in quotes. (#588)
The latest release of tiledb-cloud on PyPI has an import exception
when used with old versions of urllib3, which don’t include the
BaseHTTPResponse class. We only use that in an annotation, and it’s
the only thing from recent urllib3 that we use, so by putting it
in quotes, we can avoid evaluating it and triggering the error.