Validate "memory" string in given resources.
It's easy to forget that the memory string requires a unit suffix to
specify something other than bytes.  That results in users passing "2",
thinking that it's Gigabytes but it's really bytes.

This commit validates the given memory string. If it's a unitless number
less than 10,000,000 (i.e., 10 Mb), the code raises an exception.