Allow passing URIs for asset-related functions. (#627)
We had a mix of functions that accepted a (namespace, name) pair,
and functions which accepted a URI. This is annoying. In this change,
we update all these functions to also accept URIs.

In most cases, the namespace and name are positional arguments,
and often have other positional arguments after them. This means
we usually can't just accept a URI in the same positional argument,
because that would end up with everything else offset.

Instead, in these cases, we add a new param, `dest_uri`, which a user
can pass instead of the namespace/name pair. Support for namespace/name
is maintained, and we ensure that both are never passed.