Add datetime parsing for executing UDFs at timestamps.
This change adds the initial step for executing UDFs at timestamps.
To allow users to specify a timestamp of a registered UDF to use,
we add `@` followed by a timestamp to the name of the UDF.

For example, to execute the version of `my/udf` as it was as of
2023-06-17 at 07:16 UTC, they would execute:

    "my/udf@2023-06-17 07:16"

This is unambiguous and avoids the need to add a new parameter, which
would be difficult to name and could shadow parameters in user code.